+++ src/utils/file.js
@@ -5,8 +5,8 @@
  */
 
 // regexp to deal with file name & path
-const REMOVE_PATH = /^.*(\\|\/|\:)/;
-const REMOVE_EXT = /\.[^\.]*$/;
+const PATH = /^.*(\\|\/|\:)/;
+const EXT = /\.[^\.]*$/;
 
 
 /**
@@ -14,11 +14,23 @@ const REMOVE_EXT = /\.[^\.]*$/;
  * @public
  * @memberof utils.file
  * @name getBasename
- * @param  {string} path- -  path containing the filename
+ * @param  {string} path -  path containing the basename to extract
  * @returns {string} the base name without path information.
  */
 export function getBasename(path) {
-    return path.replace(REMOVE_PATH, "").replace(REMOVE_EXT, "");
+    return path.replace(PATH, "").replace(EXT, "");
+}
+
+/**
+ * return the path of the file
+ * @public
+ * @memberof utils.file
+ * @name getPath
+ * @param  {string} path - the copmplete file path to extract the path from
+ * @returns {string} the extracted path
+ */
+export function getPath(path) {
+    return path.match(PATH)[0];
 }
 
 /**
@@ -26,7 +38,7 @@ export function getBasename(path) {
  * @public
  * @memberof utils.file
  * @name getExtension
- * @param  {string} path- -  path containing the filename
+ * @param  {string} path - path containing the filename and extension to extract
  * @returns {string} filename extension.
  */
 export function getExtension(path) {